src/cspeed_tools/cogs/images_cog.py
from discord.ext import commands

class ImagesCog(commands.Cog):
    def __init__(self, bot: commands.Bot):
        self.bot = bot

async def setup(bot: commands.Bot) -> None:
    await bot.add_cog(ImagesCog(bot))
