pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cspeed-tools"
version = "1.0.0"
description = "c.speed-tools Discord bot"
requires-python = ">=3.11"
dependencies = [
  "discord.py==2.6.3",
  "aiohttp>=3.9",
  "aiosqlite>=0.20",
  "asyncpg>=0.29",
  "pydantic-settings>=2.2",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "ruff>=0.6", "black>=24"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
